+++ packages/markets/src/deprecated/favoritesList/favoritesList.script.ts
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useState } from "react";
 import { MarketsType, useMarkets } from "@orderly.network/hooks";
+import { useMarketsContext } from "../../components/marketsProvider";
 import { searchBySymbol, useSort } from "../../utils";
-import { useMarketsContext } from "../marketsProvider";
 
 export type UseFavoritesListReturn = ReturnType<typeof useFavoritesListScript>;
 
@@ -20,7 +20,7 @@ export const useFavoritesListScript = () => {
       ?.filter(
         (item) =>
           item.tabs?.findIndex((tab) => tab.id === selectedFavoriteTab.id) !==
-          -1
+          -1,
       )
       ?.map((fav) => {
         const index = data?.findIndex((item) => item.symbol === fav.name);