+++ packages/markets/src/deprecated/newListingList/widget.tsx
@@ -1,7 +1,7 @@
-import { NewListingList } from "./newListingList.ui";
-import { useNewListingListScript } from "./newListingList.script";
-import { GetColumns } from "../../type";
 import { DataTableClassNames } from "@orderly.network/ui";
+import { GetColumns } from "../../type";
+import { useNewListingListScript } from "./newListingList.script";
+import { NewListingList } from "./newListingList.ui";
 
 export type NewListingListWidgetProps = {
   getColumns?: GetColumns;
@@ -11,7 +11,7 @@ export type NewListingListWidgetProps = {
 };
 
 export const NewListingListWidget: React.FC<NewListingListWidgetProps> = (
-  props
+  props,
 ) => {
   const state = useNewListingListScript();
   return <NewListingList {...state} {...props} />;