+++ packages/wallet-connector/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @orderly.network/wallet-connector
 
+## 2.1.2
+
+### Patch Changes
+
+- publish
+- 7874fa6dd: publish
+- Updated dependencies
+- Updated dependencies [7874fa6dd]
+  - @orderly.network/core@2.1.2
+  - @orderly.network/hooks@2.1.2
+  - @orderly.network/types@2.1.2
+  - @orderly.network/ui@2.1.2
+  - @orderly.network/utils@2.1.2
+
+## 2.1.2-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/core@2.1.2-alpha.0
+  - @orderly.network/hooks@2.1.2-alpha.0
+  - @orderly.network/types@2.1.2-alpha.0
+  - @orderly.network/ui@2.1.2-alpha.0
+  - @orderly.network/utils@2.1.2-alpha.0
+
 ## 2.1.1
 
 ### Patch Changes