+++ packages/walletconnect/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @orderly.network/web3-modal
 
+## 2.1.2
+
+### Patch Changes
+
+- publish
+- 7874fa6dd: publish
+- Updated dependencies
+- Updated dependencies [7874fa6dd]
+  - @orderly.network/hooks@2.1.2
+
+## 2.1.2-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/hooks@2.1.2-alpha.0
+
 ## 2.1.1
 
 ### Patch Changes